# synkdata/docs/api.yaml
openapi: 3.0.3
info:
  title: SynkData Identity Intelligence Platform API
  description: |
    Plataforma de inteligencia de identidad para verificación, antecedentes,
    cumplimiento normativo y análisis de riesgo enfocada en México y LATAM.

    ## Autenticación
    Todos los endpoints protegidos requieren un token JWT en el header
    `Authorization: Bearer <token>`.

    ## Códigos de respuesta
    - `200` Operación exitosa
    - `201` Recurso creado
    - `400` Error de validación en los datos de entrada
    - `401` Token de autenticación inválido o expirado
    - `403` Permisos insuficientes para la operación
    - `404` Recurso no encontrado
    - `429` Límite de peticiones excedido
    - `500` Error interno del servidor
  version: 1.0.0
  contact:
    name: SynkData Team
    email: dev@synkdata.io
    url: https://synkdata.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Servidor de desarrollo
  - url: https://api.synkdata.io/api/v1
    description: Servidor de producción
security:
  - BearerAuth: []
...
paths:
...
  /screening/person:
    post:
      tags: [Screening]
      summary: Screening de persona física contra listas restrictivas
      description: |
        Realiza el screening completo de una persona física contra todas
        las listas restrictivas y de compliance configuradas: OFAC, ONU,
        OpenSanctions, Interpol, SAT 69-B, DOF, SCJN y PEP.
      operationId: screenPerson
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreeningPersonRequest'
            example:
              name: "JUAN PEREZ LOPEZ"
              curp: "PELJ800101HDFRRN09"
              rfc: "PELJ800101XYZ"
              country: "MX"
              threshold: 0.85
      responses:
        '200':
          description: Screening completado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreeningResult'
  /screening/entity:
    post:
      tags: [Screening]
      summary: Screening de persona moral o entidad corporativa
      description: |
        Realiza el screening de una persona moral o entidad corporativa
        contra las listas restrictivas aplicables.
      operationId: screenEntity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreeningEntityRequest'
            example:
              name: "Corporativo Alfa S.A. de C.V."
              country: "MX"
      responses:
        '200':
          description: Screening completado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreeningResult'
  /screening/sat-69b:
    post:
      tags: [Screening]
      summary: Verificación en lista SAT 69-B
      description: |
        Verifica si un contribuyente está listado en el artículo 69-B
        del Código Fiscal de la Federación (presunción de operaciones simuladas).
      operationId: screenSat69b
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sat69bRequest'
            example:
              name: "Empresa Ejemplo S.A. de C.V."
              rfc: "EEE850101ABC"
      responses:
        '200':
          description: Verificación completada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sat69bResult'
  /screening/pep:
    post:
      tags: [Screening]
      summary: Verificación de Persona Políticamente Expuesta (PEP)
      operationId: screenPep
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 500
                  description: Nombre de la persona a verificar como PEP
                  example: "ANDRÉS MANUEL LÓPEZ OBRADOR"
                country:
                  type: string
                  maxLength: 5
                  default: "MX"
                  description: País para contextualizar la búsqueda PEP
      responses:
        '200':
          description: Verificación completada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PepResult'
  /screening/{request_id}:
    get:
      tags: [Screening]
      summary: Consultar resultado de screening previo
      operationId: getScreeningResult
      security:
        - BearerAuth: []
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Identificador único de la solicitud
      responses:
        '200':
          description: Resultado encontrado
        '404':
          description: Solicitud no encontrada

  # ── IDENTITY ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint de autenticación.
        Formato: `Authorization: Bearer <token>`
  schemas:
...
    ScreeningPersonRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 500
          description: Nombre completo de la persona a screening
          example: "JUAN PEREZ LOPEZ"
        curp:
          type: string
          maxLength: 18
          description: CURP de la persona (opcional)
          example: "PELJ800101HDFRRN09"
        rfc:
          type: string
          maxLength: 13
          description: RFC de la persona (opcional)
          example: "PELJ800101XYZ"
        country:
          type: string
          maxLength: 5
          default: "MX"
          description: País de nacionalidad (código ISO)
          example: "MX"
        threshold:
          type: number
          minimum: 0.5
          maximum: 1.0
          description: Umbral de similitud personalizado
          example: 0.85
    ScreeningEntityRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 500
          description: Nombre de la entidad corporativa
          example: "Corporativo Alfa S.A. de C.V."
        country:
          type: string
          maxLength: 5
          default: "MX"
          description: País de registro
        threshold:
          type: number
          minimum: 0.5
          maximum: 1.0
          description: Umbral de similitud personalizado
    Sat69bRequest:
      type: object
      required: [name, rfc]
      properties:
        name:
          type: string
          description: Nombre del contribuyente
          example: "Empresa Ejemplo S.A. de C.V."
        rfc:
          type: string
          minLength: 12
          maxLength: 13
          description: RFC del contribuyente
          example: "EEE850101ABC"
...
    ScreeningResult:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchDetail'
        total_hits:
          type: integer
          minimum: 0
        max_score:
          type: number
          minimum: 0.0
          maximum: 1.0
        risk_level:
          type: string
          enum: [none, low, medium, high, critical]
        sources_checked:
          type: array
          items:
            type: string
        sources_failed:
          type: array
          items:
            type: string
        timestamp:
          type: string
    MatchDetail:
      type: object
      properties:
        source:
          type: string
          description: Fuente donde se encontró la coincidencia
          example: "ofac"
        score:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Puntuación de similitud
          example: 0.95
        match_type:
          type: string
          enum: [exact, fuzzy, phonetic, alias]
          description: Tipo de coincidencia
        entity_name:
          type: string
          description: Nombre de la entidad coincidente
          example: "JUAN GOMEZ RODRIGUEZ"
        entity_data:
          type: object
          additionalProperties: true
          description: Datos adicionales de la entidad
        is_confirmed:
          type: boolean
          description: Si la coincidencia fue confirmada manualmente
    Sat69bResult:
      type: object
      properties:
        is_listed:
          type: boolean
        rfc:
          type: string
        company_name:
          type: string
        status:
          type: string
          description: "Estado: presunto, desvirtuado, definitivo"
        publication_date:
          type: string
        docket_number:
          type: string
        observations:
          type: string
    PepResult:
      type: object
      properties:
        is_pep:
          type: boolean
        positions:
          type: array
          items:
            type: string
        country:
          type: string
        level:
          type: string
          description: "Nivel: national, state, municipal"
        source:
          type: string
...